// app/lib/type/decoration.type.ts
export type ThemeOption = "none" | "primary" | "white" | "slate" | "black";
export type DarkThemeOption = "none" | "slate" | "transparent";

export type SizeOption = "xs" | "sm" | "md" | "lg" | "xl";
export type RadiusOption =
  | "none"
  | "default"
  | "sm"
  | "md"
  | "lg"
  | "xl"
  | "full";
export type FontWeightOption = "thin" | "light" | "normal" | "medium" | "bold";
export type ShadowOption = "none" | "default" | "sm" | "md" | "lg" | "xl";
export type BorderWidthOption = "none" | "0" | "1" | "2" | "4";
export type BorderColorOption = "none" | "transparent" | "slate" | "black";
export type AlignOption = "left" | "center" | "right" | "between";
export type ModeSelectedOption = "icon" | "background" | "none";